requirements.txt
absl-py==0.11.0
astor==0.8.1
bleach==1.5.0
certifi==2020.6.20
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.35.0
h5py==2.10.0
html5lib==0.9999999
imagecodecs==2019.12.31
imagecodecs-lite==2020.1.31
imageio==2.9.0
importlib-metadata==2.1.1
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.0.3
networkx==2.4
numpy==1.16.6
opencv-python==4.4.0.42
Pillow==7.2.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.15.0
scipy==1.4.1
six==1.15.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
tifffile==2019.7.26.2
tqdm==4.61.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
